pyproject.toml
[project]
name            = "pilottai-tools"
version         = "0.1.0"
description     = "Add your description here"
readme          = "README.md"
requires-python = ">=3.10"
dependencies    = [
    "pydantic>=2.11.7",
    "python-docx>=1.2.0",
    "beautifulsoup4>=4.13.4",
    "pillow>=11.3.0",
    "redis>=6.2.0",
    "aiohttp>=3.12.14",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "markdown>=3.8.2",
    "pyyaml>=6.0.2",
    "opencv-python>=4.12.0.88",
    "neo4j>=5.28.1",
    "pilottai>=0.2.7.36",
]

[project.optional-dependencies]
docs        = ["python-docx", "redis", "mkdocs", "mkdocs-material", "black", "ruff", "pytest", "pre-commit"]
dev         = ["black", "ruff", "pytest", "pre-commit"]
image       = ["pillow>=11.3.0", "numpy>=2.2.6", "opencv-python>=4.12.0.88",]
web         = ["beautifulsoup4>=4.13.4", "aiohttp>=3.12.14",]
knowledge   = ["numpy>=2.2.6", "pillow>=11.3.0", "beautifulsoup4>=4.13.4", "aiohttp>=3.12.14", "pandas>=2.3.1",
               "markdown>=3.8.2", "pyyaml>=6.0.2", "opencv-python>=4.12.0.88",] # add here
short       = ["redis>=6.2.0"]
long        = ["neo4j>=5.28.1"]
redis       = ["redis>=6.2.0",]
graph       = ["neo4j>=5.28.1"]
memory      = ["redis>=6.2.0", "numpy>=2.2.6", "neo4j>=5.28.1",] # add here
all         = ["redis>=6.2.0", "pillow>=11.3.0", "beautifulsoup4>=4.13.4", "aiohttp>=3.12.14", "numpy>=2.2.6", "pandas>=2.3.1",
               "markdown>=3.8.2", "pyyaml>=6.0.2", "opencv-python>=4.12.0.88", "neo4j>=5.28.1",] # add here


[build-system]
requires        = ["hatchling"]
build-backend   = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
dependencies = [
    "black",
    "ruff",
    "pytest",
    "pre-commit"
]

[tool.hatch.envs.docs]
dependencies = [
    "mkdocs",
    "mkdocs-material"
]
